requirements.txt
PyQt6>=6.6.0
opencv-python>=4.9.0
insightface>=0.7.3
onnxruntime==1.20.1
numpy>=1.26.0
httpx>=0.27.0
PyJWT>=2.8.0
Pillow>=10.2.0
pygrabber>=0.2
